requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
openai>=1.16.0
python-dotenv==1.0.0
pyannote.audio==3.1.1
numpy==1.26.4 